+++ ModDebuggingMod/Card.cs
@@ -15,7 +15,7 @@ namespace ModDebuggingMod
         {
             CardInfo info = CardHelper.NewCard(true, "wstl", "wstlcard", "Debug",
                 attack: 1, health: 1, blood: 0, bones: 0, energy: 0, gems: null)
-                .AddAbilities(Ability.RandomAbility)
+                .AddAbilities(Sporogenic.ability)
                 .SetPortraits("Untitled", pixelPortraitName: "buffBell.png");
         }
     }